# openapi-schema.yml
openapi: 3.0.3
info:
  title: Dognews Server API
  version: 1.0.0
  description: Dognews Server client API
paths:
...
  /submissions:
    get:
      operationId: submissions_list
      description: |-
        Submitted articles for review

        **Permission restrictions:**
        + `IsAuthenticated`: *Rejects all operations if the user is not authenticated*
        + `IsOwnerOrStaff`: *Blocks update/partial_updated/destroy if:     * the user is NOT in the staff group     * AND if the model has a property called 'owner' and its value differs from the request user     Everything else is allowed*
        + `DjangoModelPermissions`: *The request is authenticated using `django.contrib.auth` permissions.     See: https://docs.djangoproject.com/en/dev/topics/auth/#permissions      It ensures that the user is authenticated, and has the appropriate     `add`/`change`/`delete` permissions on the model.      This permission can only be applied against view classes that     provide a `.queryset` attribute.*
      parameters:
      - in: query
        name: analysis__status
        schema:
          type: string
          enum:
          - failed
          - passed
          - pending
      - in: query
        name: analysis__status__isnull
        schema:
          type: boolean
      - in: query
        name: fetch__isnull
        schema:
          type: boolean
      - in: query
        name: fetch__status
        schema:
          type: string
          enum:
          - fetched
          - pending
          - rej_error
          - rej_fetch
      - in: query
        name: fetch__status__isnull
        schema:
          type: boolean
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - in: query
        name: moderation__isnull
        schema:
          type: boolean
      - in: query
        name: moderation__status
        schema:
          type: string
          enum:
          - accepted
          - pending
          - rejected
      - in: query
        name: moderation__status__isnull
        schema:
          type: boolean
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - in: query
        name: status
        schema:
          type: string
          enum:
          - accepted
          - pending
          - rej_banned
          - rej_fetch
          - rej_mod
          - rej_sentim
      tags:
      - submissions
      security:
      - basicAuth: []
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSubmissionList'
          description: ''
    post:
      operationId: submissions_create
      description: |-
        Submitted articles for review

        **Permission restrictions:**
        + `IsAuthenticated`: *Rejects all operations if the user is not authenticated*
        + `IsOwnerOrStaff`: *Blocks update/partial_updated/destroy if:     * the user is NOT in the staff group     * AND if the model has a property called 'owner' and its value differs from the request user     Everything else is allowed*
        + `DjangoModelPermissions`: *The request is authenticated using `django.contrib.auth` permissions.     See: https://docs.djangoproject.com/en/dev/topics/auth/#permissions      It ensures that the user is authenticated, and has the appropriate     `add`/`change`/`delete` permissions on the model.      This permission can only be applied against view classes that     provide a `.queryset` attribute.*
      tags:
      - submissions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmissionRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SubmissionRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SubmissionRequest'
        required: true
      security:
      - basicAuth: []
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submission'
          description: ''
  /submissions/{id}:
    get:
      operationId: submissions_retrieve
      description: |-
        Submitted articles for review

        **Permission restrictions:**
        + `IsAuthenticated`: *Rejects all operations if the user is not authenticated*
        + `IsOwnerOrStaff`: *Blocks update/partial_updated/destroy if:     * the user is NOT in the staff group     * AND if the model has a property called 'owner' and its value differs from the request user     Everything else is allowed*
        + `DjangoModelPermissions`: *The request is authenticated using `django.contrib.auth` permissions.     See: https://docs.djangoproject.com/en/dev/topics/auth/#permissions      It ensures that the user is authenticated, and has the appropriate     `add`/`change`/`delete` permissions on the model.      This permission can only be applied against view classes that     provide a `.queryset` attribute.*
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this submission.
        required: true
      tags:
      - submissions
      security:
      - basicAuth: []
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submission'
          description: ''
    put:
      operationId: submissions_update
      description: |-
        Submitted articles for review

        **Permission restrictions:**
        + `IsAuthenticated`: *Rejects all operations if the user is not authenticated*
        + `IsOwnerOrStaff`: *Blocks update/partial_updated/destroy if:     * the user is NOT in the staff group     * AND if the model has a property called 'owner' and its value differs from the request user     Everything else is allowed*
        + `DjangoModelPermissions`: *The request is authenticated using `django.contrib.auth` permissions.     See: https://docs.djangoproject.com/en/dev/topics/auth/#permissions      It ensures that the user is authenticated, and has the appropriate     `add`/`change`/`delete` permissions on the model.      This permission can only be applied against view classes that     provide a `.queryset` attribute.*
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this submission.
        required: true
      tags:
      - submissions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmissionRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SubmissionRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SubmissionRequest'
        required: true
      security:
      - basicAuth: []
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submission'
          description: ''
    patch:
      operationId: submissions_partial_update
      description: |-
        Submitted articles for review

        **Permission restrictions:**
        + `IsAuthenticated`: *Rejects all operations if the user is not authenticated*
        + `IsOwnerOrStaff`: *Blocks update/partial_updated/destroy if:     * the user is NOT in the staff group     * AND if the model has a property called 'owner' and its value differs from the request user     Everything else is allowed*
        + `DjangoModelPermissions`: *The request is authenticated using `django.contrib.auth` permissions.     See: https://docs.djangoproject.com/en/dev/topics/auth/#permissions      It ensures that the user is authenticated, and has the appropriate     `add`/`change`/`delete` permissions on the model.      This permission can only be applied against view classes that     provide a `.queryset` attribute.*
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this submission.
        required: true
      tags:
      - submissions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedSubmissionRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedSubmissionRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedSubmissionRequest'
      security:
      - basicAuth: []
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submission'
          description: ''
    delete:
      operationId: submissions_destroy
      description: |-
        Submitted articles for review

        **Permission restrictions:**
        + `IsAuthenticated`: *Rejects all operations if the user is not authenticated*
        + `IsOwnerOrStaff`: *Blocks update/partial_updated/destroy if:     * the user is NOT in the staff group     * AND if the model has a property called 'owner' and its value differs from the request user     Everything else is allowed*
        + `DjangoModelPermissions`: *The request is authenticated using `django.contrib.auth` permissions.     See: https://docs.djangoproject.com/en/dev/topics/auth/#permissions      It ensures that the user is authenticated, and has the appropriate     `add`/`change`/`delete` permissions on the model.      This permission can only be applied against view classes that     provide a `.queryset` attribute.*
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this submission.
        required: true
      tags:
      - submissions
      security:
      - basicAuth: []
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
  /submissions/{submission_pk}/votes:
    get:
      operationId: submissions_votes_list
      description: |-
        Vote management /submissions/(id)/votes (get, post)

        **Permission restrictions:**
        + `IsAuthenticated`: *Rejects all operations if the user is not authenticated*
        + `IsOwnerOrModeratorOrStaff`: *Blocks update/partial_updated/destroy if:     * the user is NOT in the staff group     * AND if the model has a property called 'owner' and its value differs from the request user     * AND if the user is not in the Moderators group     Everything else is allowed*
        + `DjangoModelPermissions`: *The request is authenticated using `django.contrib.auth` permissions.     See: https://docs.djangoproject.com/en/dev/topics/auth/#permissions      It ensures that the user is authenticated, and has the appropriate     `add`/`change`/`delete` permissions on the model.      This permission can only be applied against view classes that     provide a `.queryset` attribute.*
      parameters:
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      - in: path
        name: submission_pk
        schema:
          type: string
          pattern: ^\d+$
        required: true
      tags:
      - submissions
      security:
      - basicAuth: []
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedVoteList'
          description: ''
    post:
      operationId: submissions_votes_create
      description: |-
        Vote management /submissions/(id)/votes (get, post)

        **Permission restrictions:**
        + `IsAuthenticated`: *Rejects all operations if the user is not authenticated*
        + `IsOwnerOrModeratorOrStaff`: *Blocks update/partial_updated/destroy if:     * the user is NOT in the staff group     * AND if the model has a property called 'owner' and its value differs from the request user     * AND if the user is not in the Moderators group     Everything else is allowed*
        + `DjangoModelPermissions`: *The request is authenticated using `django.contrib.auth` permissions.     See: https://docs.djangoproject.com/en/dev/topics/auth/#permissions      It ensures that the user is authenticated, and has the appropriate     `add`/`change`/`delete` permissions on the model.      This permission can only be applied against view classes that     provide a `.queryset` attribute.*
      parameters:
      - in: path
        name: submission_pk
        schema:
          type: string
          pattern: ^\d+$
        required: true
      tags:
      - submissions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoteRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/VoteRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/VoteRequest'
      security:
      - basicAuth: []
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vote'
          description: ''
...
components:
  schemas:
...
    PaginatedSubmissionList:
      type: object
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=400&limit=100
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=200&limit=100
        results:
          type: array
          items:
            $ref: '#/components/schemas/Submission'
...
    PaginatedVoteList:
      type: object
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=400&limit=100
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=200&limit=100
        results:
          type: array
          items:
            $ref: '#/components/schemas/Vote'
...
    PatchedSubmissionRequest:
      type: object
      description: A submission object that is in initial processing
      properties:
        target_url:
          type: string
          format: uri
          minLength: 1
          maxLength: 200
        title:
          type: string
          maxLength: 120
        description:
          type: string
          maxLength: 250
        date:
          type: string
          format: date-time
          nullable: true
        fetch:
          allOf:
          - $ref: '#/components/schemas/FetchRequest'
          nullable: true
        moderation:
          allOf:
          - $ref: '#/components/schemas/ModerationRequest'
          nullable: true
        votes:
          type: array
          items:
            $ref: '#/components/schemas/VoteRequest'
          nullable: true
    Submission:
      type: object
      description: A submission object that is in initial processing
      properties:
        id:
          type: integer
          readOnly: true
        url:
          type: string
          format: uri
          readOnly: true
        target_url:
          type: string
          format: uri
          maxLength: 200
        status:
          type: string
          readOnly: true
        owner:
          type: string
          format: uri
          readOnly: true
        title:
          type: string
          maxLength: 120
        description:
          type: string
          maxLength: 250
        date:
          type: string
          format: date-time
          nullable: true
        fetch:
          allOf:
          - $ref: '#/components/schemas/Fetch'
          nullable: true
        moderation:
          allOf:
          - $ref: '#/components/schemas/Moderation'
          nullable: true
        votes:
          type: array
          items:
            $ref: '#/components/schemas/Vote'
          nullable: true
      required:
      - id
      - owner
      - status
      - target_url
      - url
    SubmissionRequest:
      type: object
      description: A submission object that is in initial processing
      properties:
        target_url:
          type: string
          format: uri
          minLength: 1
          maxLength: 200
        title:
          type: string
          maxLength: 120
        description:
          type: string
          maxLength: 250
        date:
          type: string
          format: date-time
          nullable: true
        fetch:
          allOf:
          - $ref: '#/components/schemas/FetchRequest'
          nullable: true
        moderation:
          allOf:
          - $ref: '#/components/schemas/ModerationRequest'
          nullable: true
        votes:
          type: array
          items:
            $ref: '#/components/schemas/VoteRequest'
          nullable: true
      required:
      - target_url
...
    ValueEnum:
      enum:
      - 1
      - -1
      - -100
      type: integer
    Vote:
      type: object
      description: |-
        Votes are provided in Lists and don't link back to their
        submissions once serialized
      properties:
        value:
          allOf:
          - $ref: '#/components/schemas/ValueEnum'
          minimum: -32768
          maximum: 32767
    VoteRequest:
      type: object
      description: |-
        Votes are provided in Lists and don't link back to their
        submissions once serialized
      properties:
        value:
          allOf:
          - $ref: '#/components/schemas/ValueEnum'
          minimum: -32768
          maximum: 32767
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT